requirements.txt
faker
rich
flask
pyinstaller
anthropic
openai
google-generativeai
xai
requests
python-docx

